requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.79.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==41.0.7
python-multipart==0.0.6